Make it possible for Lenny to be sad
We all know that even Lenny can be sad from time to time. So it was
essential for the script to allow Lenny to represent his emotions more
accurately.

From now on the user can specify the `--sad` parameter in order to
receive a sad Lenny instead of a happy Lenny.

Example:
    $ lennyface --sad
    $ xsel --clipboard
      ( ͡° ʖ̯ ͡°)